MY-Ground/consumer_two/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
pika==1.3.1
protobuf==3.20.3
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0